% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2k-package.R, R/r2k.R
\docType{package}
\name{r2k}
\alias{r2k}
\alias{r2k-package}
\title{R Functions Modified to Immediately Return Kable Outputs}
\usage{
r2k(x)
}
\arguments{
\item{x}{Object to convert to \code{\link[knitr]{kable}}.}
}
\value{
\code{\link[knitr]{kable}}
}
\description{
For example, table_r2k(x, y) is like table(x, y) but produces a kable output
that prints neatly to the RStudio Viewer or to an R Markdown document. Other
functions include lm_r2k, glm_r2k, gee_r2k, coxph_r2k, and svyglm_r2k. Also
includes a generic r2k function for converting R objects to kables.

Calls \code{\link[knitr]{kable}} and \code{\link[kableExtra]{kable_styling}}
as appropriate based on class of input object and creates object ready to
snip or copy/paste out of R. Currently works for data frames and matrices.
}
\details{
\tabular{ll}{
Package: \tab r2k \cr
Type: \tab Package \cr
Version: \tab 1.1.1 \cr
Date: \tab 2020-03-28 \cr
License: \tab GPL-3 \cr
}
}
\examples{
r2k(mtcars)


}
\author{
Dane R. Van Domelen \cr \email{vandomed@gmail.com}
}
