% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_k.R
\name{table_k}
\alias{table_k}
\title{Alternate Version of 'table' that Returns a Formatted Kable}
\usage{
table_k(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link[base]{table}}.}
}
\value{
\code{\link[knitr]{kable}} object.
}
\description{
Works just like \code{\link[base]{table}}, but returns a formatted
\code{\link[knitr]{kable}} which should print nicely in a report.
}
\examples{
# Frequency of mtcars$cyl
table_v(mtcars$cyl)

# Crosstab of cyl and gear
table_v(mtcars$cyl, mtcars$gear)

# Crosstab of cyl and gear, across strata of am
table_v(mtcars$cyl, mtcars$gear, mtcars$am)


}
