% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_r2k.R
\name{table_r2k}
\alias{table_r2k}
\title{Alternate Version of 'table' that Returns a Neatly Formatted Kable}
\usage{
table_r2k(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link[base]{table}}.}
}
\value{
\code{\link[knitr]{kable}}
}
\description{
Calls \code{\link[base]{table}}, does some basic formatting, and then calls
\code{\link[knitr]{kable}} and \code{\link[kableExtra]{kable_styling}}.
}
\examples{
# Crosstab of gears and cylinder using mtcars dataset
table_r2k(mtcars$cyl, mtcars$gear)


}
