% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_v.R
\name{table_v}
\alias{table_v}
\title{Alternate Version of 'table' that Also Prints to RStudio Viewer}
\usage{
table_v(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link[base]{table}}.}
}
\value{
\code{\link[base]{table}} object.
}
\description{
Works just like \code{\link[base]{table}}, but prints a formatted version of
the table to the Viewer.
}
\examples{
# Frequency of mtcars$cyl
table_v(mtcars$cyl)

# Crosstab of cyl and gear
table_v(mtcars$cyl, mtcars$gear)

# Crosstab of cyl and gear, across strata of am
table_v(mtcars$cyl, mtcars$gear, mtcars$am)


}
