% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2k.R
\name{r2k}
\alias{r2k}
\title{Shortcut to Convert an R Object to a Kable}
\usage{
r2k(x)
}
\arguments{
\item{x}{Object to convert to \code{\link[knitr]{kable}}.}
}
\value{
\code{\link[knitr]{kable}}
}
\description{
Calls \code{\link[knitr]{kable}} and \code{\link[kableExtra]{kable_styling}}
as appropriate based on class of input object and creates object ready to
snip or copy/paste out of R. Currently works for data frames and matrices.
}
\examples{
r2k(mtcars)


}
